Complete initial display compatibility
Display now turns on and draws shapes and lines in the same way as it
did on the Teensy. The problem was a missing break in a switch
statement that made digitalWrite() fall through to always turning pins
off.

While shapes are working, text is not. Right now, any text printed to
screen appears mirrored and rotated counterclockwise 90 degrees. Also,
object drawing appears to be slow.

The extra digitalWrite() function in main.cpp is because the version of
digitalWrite that the display driver uses by default only takes two
arguments, requiring that the GPIOx header already be set. The external
status LED is on GPIOA, whereas cs and rst for the display are GPIOC.