Add hardware wait polling and more font options
This fixes gh-2, adding a hardware wait poll on RA8875's WAIT
pin before a read or write eliminates the need to have a built
in delay.

Also, referencing gh-1, font rotation support from the RA8875 hardware
was added; however, this has not resolved the issue. Might be a deeper
problem with font rendering as a result of hardware differences between
STM and Arduino.