Clean up touch screen operation, debounce buttons
Returned back to automatic mode, as the RA8875 seems to be finicky and
the touch screen has now started working again with default settings.
Created helper functions to simplify management of touch data.

Added small debouncing statements in button click detection to prevent
double-presses on the settings and cancel buttons.