Abstract away pin handling and GPIO initialization
Create a System.h file containing a pin class abstracting away dealing
with multiple GPIO devices or initializing digital GPIO pins. Still in
the works: analog pins, SPI.

Also creating an Arduino_Compat.h file that serves as a layer between
Adafruit libraries and the STM32F4 peripheral library. Goal of this
file is to minimize the amount of changes that are required to get the
Adafruit libraries working on this MCU.

Finally, outlined goals of the refactor in a separate markdown file.