Improve GPIO, SPI syntax; restore old display libs
Reworked constructors for GPIO and SPI classes so that they can be
instantiated by referring to the name in the datasheet. Also, restored
the adafruit libs to as much of the original code as possible. Still
working on reducing the number of modifications required of the libs.

Also, reorganizing system level code for easier access.

Finally, fixing paths in the Makefile because the old library toolchain
was actually out of date and didn't support SPI devices beyond SPI3.