Clean SPI class, add UART console, improve display
A way-too-big commit encompassing a lot of different goals.

SPI class was cleaned of CS, RST pin references (those aren't
operated by the hardware level SPI interface anyways, they're manually
triggered).

Timer class was modified, losing TIM1 and gaining TIM4. TIM1 didn't
conform to the same hardware interface as the other timers, so I went
ahead and chose a timer interface that was the same as what was already
working.

Added overriding methods for writing data, commands to the display,
adding wait statements so that the system doesn't talk too quickly to
the display.

General modifications to encapsulate hardware better, and the
corresponding changes in the Makefile, etc.