Add digital lowpass filter to ECG
Added a digital lowpass filter to the ECG as well as some calibration
values to ensure that the signal displays properly on the screen. Also,
the ECG output values are being printed to the console for external
analysis.

Digital filter specs are: 10th order IIR equiripple lowpass represented
as 5 SOS in direct form II, fc = 25 Hz. The exact specifications and
cutoff frequency may change if a better filter configuration is determined.