Improve delay function, give RA8875 more init time
Working on gh-9, as it seems that many of the instabilities on boot are
related to the RA8875 chip. The RA8875 datasheet gives specific reset
time windows required for chip performance, so the code has been
adjusted to give slightly more than the minimum required time frame.

In order to do this, the delay function was modified so that the
argument passed in is in units of usec. This isn't incredibly accurate,
especially at lower delay windows because the scaling constant is a
decimal but the result has to be cast to an int. However, it gives a
fairly close estimate based on the oscilloscope readings.

Also added WAIT pin checking on all SPI transfers and added small
layout improvements to the code.