Add large fonts for numerical metrics, fix gh-13
Created large font bitmaps for the digits 0-9. Up to a three digit
number can be displayed. Also added some styling to the BPM indicator
next to the ECG readout.

Also added some registers to the RA8875 header file relating to BTE
transfers. These registers may not be necessary and might be removed
later on.