Fix gh-9, ensure ADC initialized before sampling
Resolved the issue where the device would not boot after reset when the
ECG readout was displayed. The ECG readout object was being
constructed, and the sampling ISR enabled, before the adcInit()
function could be called. Deferred enabling the ISR to an enable()
function in order to ensure a stable boot.