Fix CircleBuffer init error, add pressure pin
Fix an error in CircleBuffer where an object's CircleBuffer attribute
would be initialized instead of just being declared, then
re-initialized, causing pointer chaos and very strange graphing
outputs. Resolved by adding a `resize` function that allows a
CircleBuffer to be resized rather than require it be instantiated with
the correct size.

Also, added support for two inputs on the NIBP: pressure and pulsing
amplitude. Both signals are required to make accurate BP measurements.
Beginning to add interface infrastructure for interacting with the BP
unit as well (hence the button addition).