Adjust screen resolution, add resize functions
Changed the screen grid to a 12x12 instead of a 10x10 in order to
accommodate the UX for the NIBP unit. Also, changed the rest of the
object constructors that were causing double-deletes to the `resize()`
function introduced in 736e8ba.

Beginning work on the program logic for the NIBP test.