Use callback function for buttons, general cleanup
Add callback functions to Button objects that are called when the
Screen they are associated with checks for touch events. Also, perform
general code base maintenance, renaming, and organization of variables.

Also, update docs to reflect that I ditched the proposed Action system
in favor of the callback function because it was easier and clearer in
pretty much every conceivable way.