Change screen lifecycle to include `propogateTouch`
Update docs to describe the lifecycle hooks that occur during every
screen update. Changed `main.cpp` to more clearly reflect when these
lifecycle hooks are called.

In keeping with this, `listenForTouch` was renamed to `propogateTouch`
and the full touch screen cycle was moved inside this function. The
call sequence for a screen lifecycle iteration is now:

```
screen.update(some_delay);
screen.propogateTouch();
```

where `some_delay` is a delay period to allow the display chip to
finish rendering all updated elements.