Add sandboxes, allow changing ScreenElement values
Added a `Sandbox` ScreenElement that mounts other ScreenElement objects
inside it. Currently being used to develop the NIBP interface.

Also, added modifier functions to let the text and color of some
ScreenElement objects to be dynamically modified and updated on the
screen. Useful in the NIBP interface building process since we won't
have to recreate screen elements every time the state changes, we can
just modify the fields.