Build initial NIBP UI and state machine
Used the `Sandbox` ScreenElement to mount the UI for the NIBP module.
Hooked the display objects into `mainScreen` and build the structure
for the state machine.

Also, miscellaneous small interface fixes. Notably, removed the
distinction between a `Button` and all other `ScreenElement`s in the
definition of `Screen`. Now, a button has both an `update` method
(which runs the onClick function if it has been tapped) and an
`isTapped` method (which returns a boolean if the button was tapped).