Fix error in startup file, move under src control
Fixes gh-17 and probably many other of the transient, horrible and
impossible to debug errors. Turns out the "default" makefile for
TrueSTUDIO assumes a completely incorrect cpu and fpu in the startup.s
file. While I'm not entirely sure the impact that this had on the code,
I can tell from empirical testing that many transient issues are now
solved, including but not limited to:

- Occasional font-flipping

- Inconsistent touch screen enable

- Freeze after reset

- Inability to reflash in some branches (would require a complete
  erase before reflashing)