Enable SpO2 AFE, fix some strange merge conflicts
There are a bunch of files in this commit because of some weirdness with
transferring the git repository over to my new working environment. In
the interest of time, I didn't manually separate them out, so they can
safely be ignored (anything under `libraries/inc/Fonts/*`).

Finally enabled the SpO2 AFE chip and am now getting data back! Key
things in this commit:

1. Added the `ADC_PDN` pin to the PulseOx object to make sure the power
   stays on

2. Wrote out a basic algorithm for calculating SpO2, but it should
   definitely be refined more

3. Added a signal trace to the PulseOx object so the transmittance curve
   can be plotted alongside the ECG.

Future commits will be more granular!!! Apologies for this monstrosity.