Develop SpO2 algorithm, fix roundoff errors
The SpO2 algorithm focuses on finding the systolic and diastolic
pressure based on the mean arterial pressure (MAP). Currently, systolic
is determined as the pressure level greater than the MAP that has an RMS
amplitude of 1/3 the MAP RMS (Rms_max). Conversely, diastolic pressure
is determined as the pressure level less than the MAP that has an RMS
amplitude of 2/3 Rms_max.

This is giving pretty good result agreement with a standard at-home
automatic blood pressure cuff, but these cutoff points are still a
little crude and probably can be refined.