Fix old SpO2 commands and properly init system
Some of the legacy AFE4400 commands were being used in the startup and
pulse timing ritual, and they were causing problems on the new board.
Did some more datasheet diving and fixed some of the register values.

Also added an error indicator and self-check mechanism for the AFE4400
module. This is a really crude check, but it did the trick to help me
identify boards that were completely disconnected. I need to refine this
error handling further down the road, but that's not going to be a high
priority until after this version release.