Update interface for preclinical trials
Updated the interface to look presentable for the preclinical trials in
Nepal. The biggest change is that the ECG module was disabled on the
prototype due to hardware issues, so the ECG elements had to be removed
from the prototype.

Also added a SysTick timer for pulse calculations, implemented in the
next commit.