Refactor Console into logging system, expand USART
The logging macro borrows heavily from the design @jacobingalls
created for logging during OS development. It provides four levels of
logging as well as colored console output.

The logging macro resolves to an empty line when the program is being
compiled in production mode and a log statement when compiled in
development mode. The Makefile was updated to reflect this change:

    `make prod` --> Compile in production mode
    `make dev`  --> Compile in development mode

By default, the Makefile will compile in development mode.

Finally, the USART module was broken apart from the Console module and
expanded to cover all possible USART channels and pin combinations.